No more job success when no config found
Previously, when no configuration could be found the app crashes. There
is an error message instead now. When the task operation config could
not be found the default behaviour was to make the job successful, which
is not always the correct way. When no config is found the job is not
finished as success. Later commit will check even more correct to make
better distinction between no config and final task in config. Also
renamed the getNextTaskConfig to getCurrentTaskConfig to better reflect
what is actually going on.